+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - feat: Manage images on comments
 - feat: Manage image form display
+- feat: Add admin management panel
 
 ### Fixed
 - fix: CSS style on idea card